A few changes and much Django Confusion
Beat my head against this for a while.
Was trying to create  a bound form and validate it and then .save() it to the db. But our model form doesn't provide enough info to add anything to the db. It doesn't know about the id of the report and the user.
Gonna validate and .save() the model instead I guess.
Oh well learned some debugging tricks and changed a few things